Incident-Response-Agent/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Anthropic SDK
anthropic==0.28.0

# HTTP client (chamadas entre serviços)
httpx==0.27.0

# Rate limiting
slowapi==0.1.9

# Resilience
tenacity==8.3.0

# Observability
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-httpx==0.45b0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.20.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
hypothesis==6.100.1
